Extract dma_buffer code to separate header
